pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "darwin-annotator"
dynamic = ["version"]
description = "Fast, accurate prokaryotic genome annotation pipeline"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "FREEQUENCY1111" }]
keywords = ["bioinformatics", "genomics", "annotation", "prokaryote"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "biopython>=1.83",
    "pyhmmer>=0.10",
    "click>=8.0",
    "rich>=13.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "pydantic>=2.0",
    "pandas>=2.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pre-commit",
    "types-requests",
]
test = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-timeout>=2.2",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
darwin = "darwin.cli:main"

[project.urls]
Homepage = "https://github.com/FREEQUENCY1111/Darwin0.1"
Repository = "https://github.com/FREEQUENCY1111/Darwin0.1"

# ── Hatch (build + versioning) ───────────────────────────────
[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/darwin"]

# ── Ruff ─────────────────────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 99
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["darwin"]

# ── Mypy ─────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# ── Pytest ───────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: needs external tools installed",
    "slow: takes >30s",
]
addopts = "-ra --strict-markers"
asyncio_mode = "auto"
